Robocopy: Add lowerAlpha template function
